requirements.txt
sqlalchemy
docling

langchain
langchain-core
langchain-community
langchain-text-splitters